Removed some unneeded casting

+++ src/HidLibrary/NativeMethods.cs
@@ -16,7 +16,7 @@ namespace HidLibrary
 
 			// DEVPKEY_Device_BusReportedDeviceDesc
 			DEVPKEY_Device_BusReportedDeviceDesc.pid = 4;
-			DEVPKEY_Device_BusReportedDeviceDesc.fmtid = new Guid((uint)0x540b947e, (ushort)0x8b40, (ushort)0x45bc, 0xa8, 0xa2, 0x6a, 0x0b, 0x89, 0x4c, 0xbd, 0xa2);
+			DEVPKEY_Device_BusReportedDeviceDesc.fmtid = new Guid(0x540b947e, 0x8b40, 0x45bc, 0xa8, 0xa2, 0x6a, 0x0b, 0x89, 0x4c, 0xbd, 0xa2);
 
 			// DEVPKEY_Device_DeviceDesc
 			//DEVPKEY_Device_BusReportedDeviceDesc.pid = 2;
